bot/requirements.txt
aiogram>=3.4.0
openai>=1.14.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
python-dotenv>=1.0.0